Proyecto-4

+++ backend/routes/usuarios.js
@@ -3,20 +3,22 @@ const express = require("express");
 const router = express.Router();
 //Modulos creados
 const { Usuario } = require("../models/usuario");
+const { Proyecto } = require("../models/proyectos");
+const { Equipo } = require("../models/equipos");
 const auth = require("../middleware/auth");
 //Rutas
 //Crear un administrador
-router.post("/registrarAdmin", async(request,response)=>{
-  let admin = await Usuario.findOne({rol:'Administrador'});
-  if(admin) return response.status(400).send('Ya existe un administrador');
+router.post("/registrarAdmin", async (request, response) => {
+  let admin = await Usuario.findOne({ rol: "Administrador" });
+  if (admin) return response.status(400).send("Ya existe un administrador");
   admin = new Usuario({
-    usuario:request.body.usuario,
-    contrasenia:request.body.contrasenia,
-    rol:'Administrador'
+    usuario: request.body.usuario,
+    contrasenia: request.body.contrasenia,
+    rol: "Administrador",
   });
   let save = admin.save();
-  response.status(200).send({message:'Administrador creado con éxito'})
-})
+  response.status(200).send({ message: "Administrador creado con éxito" });
+});
 //Registrar los usuarios
 router.post("/registrarUsuarios", auth, async (request, response) => {
   //Si existe el usuario
@@ -44,7 +46,7 @@ router.post("/registrarUsuarios", auth, async (request, response) => {
     const token = usuario.generateJWT();
     response.status(200).send({ token });
   } else {
-    response.status(400).send("No tiene permisos necesarios");
+    response.status(401).send("No tiene permisos necesarios");
   }
 });
 //Listar los usuarios
@@ -58,7 +60,7 @@ router.get("/listarUsuarios", auth, async (request, response) => {
     const usuarios = await Usuario.find().sort({ rol: 1 });
     response.status(200).send(usuarios);
   } else {
-    return response.status(400).send("No tiene permisos necesarios");
+    return response.status(401).send("No tiene permisos necesarios");
   }
 });
 //Actualizar los usuarios
@@ -110,6 +112,8 @@ router.put("/actualizarUsuarios", auth, async (request, response) => {
       const save = newUsuario.save();
       response.status(200).send(newUsuario);
     }
+  } else {
+    response.status(401).send("No tiene los permisos necesarios");
   }
 });
 //Eliminar los usuarios
@@ -121,11 +125,22 @@ router.delete("/borrar/:_id", auth, async (request, response) => {
   //Si es administrador
   if (usuario.rol == "Administrador") {
     //Buscar el usuario a eliminar
-    const usuario = await Usuario.findByIdAndDelete(request.params._id);
+    const usuario = await Usuario.findById(request.params._id);
     if (!usuario) return response.status(400).send("El usuario no existe");
+    //Si pertenece a un proyecto no se elimina
+    const proyectos = await Proyecto.findOne({ idUsuario: usuario._id });
+    if (proyectos) return response.status(406).send("No se puede eliminar el usuario es creador de un proyectos");
+    //Si pertenece a un equipo
+    const equipo = await Equipo.findOne({ usuario: usuario.usuario });
+    if (equipo)
+      return response
+        .status(406)
+        .send("No se puede eliminar el usuario, pertenece a un proyecto");
+    const eliminar = await Usuario.findByIdAndDelete(request.params._id);
+    if(!eliminar) return response.status(400).send('No se pudo eliminar')
     response.status(200).send({ message: "Usuario eliminado" });
   } else {
-    response.status(400).send("No tiene los permisos necesarios");
+    response.status(401).send("No tiene los permisos necesarios");
   }
 });
